backend/requirements.txt
# Core Flask dependencies
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.5

# Database
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9

# Environment and configuration
python-dotenv==1.0.1
Werkzeug==3.0.2

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
coverage==7.3.2

# Production (optional)
gunicorn==21.2.0 